// helpers/schema.tsx
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { ColumnType } from "kysely";

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export type Timestamp = ColumnType<Date, Date | string, Date | string>;

export interface ProjectFiles {
  createdAt: Generated<Timestamp | null>;
  fileContent: string;
  filePath: string;
  fileType: string;
  id: Generated<number>;
  projectId: number;
}

export interface Projects {
  createdAt: Generated<Timestamp | null>;
  description: string | null;
  id: Generated<number>;
  prompt: string;
  title: string;
  updatedAt: Generated<Timestamp | null>;
}

export interface DB {
  projectFiles: ProjectFiles;
  projects: Projects;
}




// Note: kysely maps the table/column/enum names from snake_case to camelCase and PascalCase. When running SQL statements, make sure to use snake_case, but TypeScript code should use camelCase or PascalCase.